Move FragmentExecState to plan_fragment_executor
